rename ErrorMessage.ts file
