+++ packages/squiggle-lang/src/fr/danger.ts
@@ -20,7 +20,7 @@ import {
 } from "../library/registry/frTypes.js";
 import { FnFactory, unpackDistResult } from "../library/registry/helpers.js";
 import { ReducerContext } from "../reducer/context.js";
-import { ErrorMessage, REOther } from "../reducer/ErrorMessage.js";
+import { ErrorMessage, REOther } from "../errors.js";
 import { Lambda } from "../reducer/lambda.js";
 import * as E_A from "../utility/E_A.js";
 import { Ok, result } from "../utility/result.js";
