+++ packages/squiggle-lang/src/fr/dist.ts
@@ -14,11 +14,7 @@ import { FnFactory } from "../library/registry/helpers.js";
 import { OtherOperationError } from "../operationError.js";
 import * as Result from "../utility/result.js";
 import { Value, vDist } from "../value/index.js";
-import {
-  ErrorMessage,
-  REDistributionError,
-  REOther,
-} from "../reducer/ErrorMessage.js";
+import { ErrorMessage, REDistributionError, REOther } from "../errors.js";
 
 const CI_CONFIG = [
   { lowKey: "p5", highKey: "p95", probability: 0.9 },
