+++ packages/squiggle-lang/src/fr/scoring.ts
@@ -8,7 +8,7 @@ import {
   frRecord,
 } from "../library/registry/frTypes.js";
 import { FnFactory } from "../library/registry/helpers.js";
-import { REDistributionError, REOther } from "../reducer/ErrorMessage.js";
+import { REDistributionError, REOther } from "../errors.js";
 import * as Result from "../utility/result.js";
 import { vNumber } from "../value/index.js";
